requirements.txt
# Polymarket API client
py-clob-client

# WebSocket client for market data (CLOB WSS)
websockets==12.0

# HTTP client for market discovery and price fetching
httpx==0.27.0

# Environment configuration
python-dotenv==1.0.1

# Rich console output (optional, falls back to basic logging if not available)
rich>=13.0.0
load-dotenv